# references/openapi-11-25.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Curated For You, By You API",
    "description": "Simple backend requests",
    "version": "1.0.0"
  },
  "paths": {
    "/artists/genre": {
      "get": {
        "summary": "Get Artists By Genre",
        "description": "Returns an array of N artists based on a genre\n\nParameters\n----------\ngenre : str\n    An allowed genre that's searchable.\nn : int\n    The second of artists to return if possible.\nReturns\n-------\nlist\n    An array of N artists or\n    an array of \u003C N artists or\n    an array of zero artists",
        "operationId": "get_artists_by_genre_artists_genre_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Genre"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists/city": {
      "get": {
        "summary": "Get Artists By Genre City",
        "description": "Returns an array of N artists based on a genre and city\n\nParameters\n----------\ngenre : str\n    An allowed genre that's searchable.\ncity : str\n    An allowed city\nn : int\n    The second of artists to return if possible.\nReturns\n-------\nlist\n    An array of N artists or\n    an array of \u003C N artists or\n    an array of zero artists",
        "operationId": "get_artists_by_genre_city_artists_city_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Genre"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Get Root",
        "operationId": "get_root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/artists": {
      "get": {
        "summary": "Get Artists",
        "operationId": "get_artists_artists_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Genre"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Country"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists/{name}": {
      "get": {
        "summary": "Get Artist Info",
        "operationId": "get_artist_info_artists__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists/{name}/description": {
      "get": {
        "summary": "Get Artist Description",
        "operationId": "get_artist_description_artists__name__description_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists/{name}/image": {
      "get": {
        "summary": "Get Artist Image",
        "operationId": "get_artist_image_artists__name__image_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/artists/{name}/albums": {
      "get": {
        "summary": "Get Artist Albums",
        "operationId": "get_artist_albums_artists__name__albums_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/albums/{title}/description": {
      "get": {
        "summary": "Get Album Description",
        "operationId": "get_album_description_albums__title__description_get",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Title"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}